test: refactor error cases helpers
